+++ cli/testio.h
@@ -2,13 +2,14 @@
 #define RAWSTOR_CLI_TESTIO_H
 
 #include <stddef.h>
+#include <stdint.h>
 
 #ifdef __cplusplus
 extern "C" {
 #endif
 
 int rawstor_cli_testio(
-    unsigned int queue_size, const char* target, size_t block_size,
+    unsigned int queue_size, const char* target, uint64_t block_size,
     unsigned int count, unsigned int io_depth, int vector_mode
 );
